+++ src/consulting_copilot/__init__.py
@@ -19,12 +19,13 @@ from .conflict_triage import build_conflict_triage
 from .triage_report import build_triage_outcome_report
 from .triage_history import summarize_triage_history
 from .feedback_replay import replay_reviewer_feedback
+from .feedback_reconciliation import reconcile_feedback_with_triage
 
 __all__ = [
     "DEFAULT_THRESHOLDS", "ConsultingCopilot", "ConsultingEngagement", "DecisionScenario",
     "DecisionThresholds", "EvidenceItem", "compare_scenarios", "load_engagement", "load_scenarios",
     "normalize_interview_notes", "review_candidate_claims", "render_markdown",
     "build_evidence_lineage", "write_lineage", "build_adjudication_receipt", "validate_adjudication_receipt",
-    "build_conflict_triage", "build_triage_outcome_report", "summarize_triage_history", "replay_reviewer_feedback",
+    "build_conflict_triage", "build_triage_outcome_report", "summarize_triage_history", "replay_reviewer_feedback", "reconcile_feedback_with_triage",
 ]
 __version__ = "1.0.0"